Updated tons of commands to use the database correctly and added unified error handling of database errors

+++ commands/disapprove.js
@@ -18,6 +18,7 @@ class Disapprove extends BaseCommand {
     execute(args, message) {
         if(!message.guild) {
             message.author.send("Sorry, but this command doesn't work in direct messages!");
+            message.delete();
             return;
         }
 
@@ -31,21 +32,41 @@ class Disapprove extends BaseCommand {
 
         if(args.length >= 2) {
             let appsToDelete = [];
+            let handledMentions = 0;
             message.mentions.members.array().forEach(function(user) {
                 dbClient.getApplications(user.id, gid).then(function(userApps) {
-                    if(userApps) {
+                    if(userApps && userApps.length > 0) {
                         appsToDelete.concat(userApps);
                         user.send(message.author.username + " has disapproved your role applications!");
                     }
-                });
+                    handledMentions++;
+                    if(handledMentions === message.mentions.members.array().length) {
+                        dbClient.deleteApplications(...appsToDelete).then(function() {
+                            message.author.send("User applications were successfully disapproved!");
+                            message.delete();
+                        }).catch(errorFunc.bind(this, message));
+                    }
+                }).catch(errorFunc.bind(this, message));
             });
-            dbClient.deleteApplications(...appsToDelete);
         }
         else if(args.length === 1) {
             message.author.send('\'disapprove\' command invalid: Missing user parameter!');
+            message.delete();
         }
-        message.delete();
     }
 }
 
 module.exports = Disapprove;
+
+/**
+ * Relays an error message to the default error output and tells the user to consult admins.
+ * @param [message] {Message}
+ * @param error {Error}
+ */
+function errorFunc(message, error) {
+    logger.error(error);
+    if(message) {
+        message.author.send("Sorry, but something went wrong. If this keeps happening, please tell your admin!");
+        message.delete();
+    }
+}
